moved utilities to util folder
